add Drawful2 support
This commit adds support for Drawful2. The bot plays Drawful 2 by
searching for the prompt using Bing's image search, and rendering it
similarly to the Tee K.O. rendering. In addition, it can set a profile
picture (either selected randomly from the current directory, or a
random image from E926).

This commit was the result of a few pair-programming sessions.

Authors:
- William Osler
- Cassandra Guillott